// src/app/vampire-masquerade/model/masquerade-sheet/vampire-masquerade/disciplines/VampireVTMDisciplines.ts
import { VampireVTMDisciplineLevel } from "./VampireVTMDisciplineLevel";
import { VampireVTMMagicDisciplineLevel } from "./VampireVTMMagicDisciplineLevel";

export interface VampireVTMDisciplines {
  animalism: VampireVTMDisciplineLevel;
  auspex: VampireVTMDisciplineLevel;
  celerity: VampireVTMDisciplineLevel;
  chimerstry: VampireVTMDisciplineLevel;
  dementation: VampireVTMDisciplineLevel;
  dominate: VampireVTMDisciplineLevel;

  fortitude: VampireVTMDisciplineLevel;
  necromancy: VampireVTMMagicDisciplineLevel;
  obfuscate: VampireVTMDisciplineLevel;
  obtenebration: VampireVTMDisciplineLevel;
  potence: VampireVTMDisciplineLevel;
  presence: VampireVTMDisciplineLevel;

  protean: VampireVTMDisciplineLevel;
  quietus: VampireVTMDisciplineLevel;
  serpentis: VampireVTMDisciplineLevel;
  thaumaturgy: VampireVTMMagicDisciplineLevel;
  vicissitude: VampireVTMDisciplineLevel;
}